# perf(frequency): Use byte records instead of string records