```
Benchmark 1: ./pica-v0.25 frequency -su -H 'gnd_id,label,count' '041A/*{ (9, a) | 9? && 7 =~ "^T[bfgpsu][1z]$" }' DUMP.dat --where '002@.0 =^ "A" && 041A/*.9?' -o freq.csv
  Time (mean ± σ):     244.640 s ± 20.967 s    [User: 210.942 s, System: 18.619 s]
  Range (min … max):   229.814 s … 259.465 s    2 runs

Benchmark 2: ./pica-main frequency -su -H 'gnd_id,label,count' '041A/*{ (9, a) | 9? && 7 =~ "^T[bfgpsu][1z]$" }' DUMP.dat --where '002@.0 =^ "A" && 041A/*.9?' -o freq.csv
  Time (mean ± σ):     212.046 s ±  1.081 s    [User: 192.311 s, System: 18.450 s]
  Range (min … max):   211.281 s … 212.810 s    2 runs

Summary
  './pica-main frequency -su -H 'gnd_id,label,count' '041A/*{ (9, a) | 9? && 7 =~ "^T[bfgpsu][1z]$" }' DUMP.dat --where '002@.0 =^ "A" && 041A/*.9?' -o freq.csv' ran
    1.15 ± 0.10 times faster than './pica-v0.25 frequency -su -H 'gnd_id,label,count' '041A/*{ (9, a) | 9? && 7 =~ "^T[bfgpsu][1z]$" }' DUMP.dat --where '002@.0 =^ "A" && 041A/*.9?' -o freq.csv'
```